refactor(utils): Phase 3 - Move transform utilities to transform/ subdirectory
Organizational refactor - no functional changes:
- src/utils/transform/response-transformer.ts (author extraction, work normalization)
- src/utils/transform/transform-work.ts (work data transformation)
- src/utils/transform/normalization.ts (ISBN/string normalization)
- src/utils/transform/book-mappers.ts (Gemini CSV to BookRecord mapping)

Updated imports in 15+ files and fixed relative paths.
All 199 smoke tests passing.

Part of Phase 3: High-risk utility consolidation.
