+++ src/api-v3/index.ts
@@ -32,7 +32,7 @@ import { generateBookEmbedding, storeEmbedding } from '../services/embedding-ser
 import { enrichMultipleBooks } from '../services/enrichment'
 import type { Env } from '../types/env'
 import { isValidEnrichedBookCacheEntry } from '../utils/validation/book-validation'
-import { normalizeTitle } from '../utils/normalization'
+import { normalizeTitle } from '../utils/transform/normalization'
 import { registerDiscoveryRoutes } from './discovery'
 import {
   buildStreamUrl,
