+++ src/handlers/book-search.ts
@@ -14,7 +14,7 @@ import type { Env } from '../types/env'
 import { writeCacheMetrics } from '../utils/analytics/analytics'
 import { detectImageQuality } from '../utils/book/book-metadata'
 import { setCached } from '../utils/cache/cache'
-import { transformWorkToGoogleFormat } from '../utils/transform-work'
+import { transformWorkToGoogleFormat } from '../utils/transform/transform-work'
 
 // ============================================================================
 // Types
