+++ src/handlers/search-handlers.ts
@@ -10,7 +10,7 @@
 import * as externalApis from '../services/external-apis'
 import type { Env } from '../types/env'
 import { createErrorResponse, ErrorCodes } from '../utils/http/response-builder'
-import { transformWorkToGoogleFormat } from '../utils/transform-work'
+import { transformWorkToGoogleFormat } from '../utils/transform/transform-work'
 
 // ============================================================================
 // Types
