+++ src/utils/cache/cache-keys.ts
@@ -1,7 +1,7 @@
 // src/utils/cache-keys.ts
 
 import { PROMPT_VERSION } from '../../prompts/csv-parser-prompt.ts'
-import { normalizeISBN } from '../normalization.js'
+import { normalizeISBN } from '../transform/normalization.js'
 
 /**
  * Generate SHA-256 hash of string using Web Crypto API
